Move cli library codes to client/
