+++ client/keys/generate.go
