+++ client/keys/remove.go
